Merge 'develop' into '1.6.x' for QloApps version 1.6.1

+++ modules/hotelreservationsystem/controllers/admin/AdminHotelfeaturesController.php
@@ -41,11 +41,6 @@ class AdminHotelFeaturesController extends ModuleAdminController
             'desc' => $this->l('Add new Features'),
             'imgclass' => 'new'
         );
-
-        $this->page_header_toolbar_btn['new'] = array(
-            'href' => $this->context->link->getAdminLink('AdminAssignHotelFeatures'),
-            'desc' => $this->l('Assign Features To Hotel'),
-        );
     }
 
     public function renderView()
@@ -58,15 +53,19 @@ class AdminHotelFeaturesController extends ModuleAdminController
 
     public function renderForm()
     {
+        if (!$this->loadObject(true)) {
+            return;
+        }
+
         $smartyVars = array();
         //lang vars
         $languages = Language::getLanguages(false);
-        $currentLangId = Configuration::get('PS_LANG_DEFAULT');
+        $currentLangId = $this->default_form_language ? $this->default_form_language : Configuration::get('PS_LANG_DEFAULT');
         $currentLang = Language::getLanguage((int) $currentLangId);
         $smartyVars['languages'] = $languages;
         $smartyVars['currentLang'] = $currentLang;
         $smartyVars['ps_img_dir'] = _PS_IMG_.'l/';
-        if ($id = Tools::getValue('id')) {
+        if ($id = $this->object->id) {
             $smartyVars['edit'] = 1;
             if (Validate::isLoadedObject($objFeatures = new HotelFeatures($id))) {
                 $featureInfo = (array) $objFeatures;
@@ -148,18 +147,6 @@ class AdminHotelFeaturesController extends ModuleAdminController
                 }
             } else {
                 $this->errors[] = $this->l('Please add atleast one Child features.');
-            }
-
-            foreach ($languages as $lang) {
-                if (!trim(Tools::getValue('parent_ftr_name_'.$lang['id_lang']))) {
-                    $objHotelFeatures->name[$lang['id_lang']] = Tools::getValue(
-                        'parent_ftr_name_'.$defaultLangId
-                    );
-                } else {
-                    $objHotelFeatures->name[$lang['id_lang']] = Tools::getValue(
-                        'parent_ftr_name_'.$lang['id_lang']
-                    );
-                }
             }
 
             if (!count($this->errors)) {
@@ -297,13 +284,20 @@ class AdminHotelFeaturesController extends ModuleAdminController
 
     public function ajaxProcessDeleteFeature()
     {
-        $idFeature = Tools::getValue('feature_id');
-        $objHotelFeatures = new HotelFeatures();
-        if ($objHotelFeatures->deleteHotelFeatures($idFeature)) {
-            die('success');
+        $response = array('status' => false);
+        if ($this->tabAccess['delete']) {
+            $idFeature = Tools::getValue('feature_id');
+            $objHotelFeatures = new HotelFeatures();
+            if ($objHotelFeatures->deleteHotelFeatures($idFeature)) {
+                $response['status'] = true;
+            } else {
+                $response['msg'] = $this->l('Some error occurred while deleting feature. Please try again.');
+            }
         } else {
-            echo 0;
+            $response['msg'] = $this->l('You do not have the permission to delete this.');
         }
+
+        $this->ajaxDie(json_encode($response));
     }
 
     public function setMedia()
